Add helm chart for node feature discovery
This packages up the upstream "node feature discovery" project[1] into
a helm chart and includes it in the ISO.  The expectation is that it's
there if end-users want to use it, but it will not be installed by
default.

Rather than make a new RPM, the helm chart has been added to the
existing stx-platform-helm package.  Unlike the other charts in that
package, for now it will be included in the load (to make it available
right away). Once the helm charts from that application are properly
installed on system installation then we can consider removing the node
feature discovery helm chart from the filesystem as it will be available
from the local helm repo.

There are two primary files describing Kubernetes resources.  They both
originally came from version 0.3.0 of the upstream project but one of
them needed to be renamed to work with "helm package" which restricts
the allowable filename suffixes.

The two upstream files have been modified to support helm overrides.
This also required adding a template for the namespace, if a custom one
is specified.

[1] https://github.com/kubernetes-sigs/node-feature-discovery/

Change-Id: I6e7c8a629994ad4da3834cbefccb94cd01259cc5
Story: 2005193
Task: 29954
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>