Fixing special conf chars on SQL_CONNECTION
The awk used to split the fm.conf file didn't takes into consideration
that "=" can be used as a special char on the database's password, and
this happens with randomly generated passwords for test. This commits
fixes that by getting the whole string after the config key after
sed'ing out the key.

Closes-Bug: #1916083

Signed-off-by: Thiago Brito <tbrito@daitan.com>
Change-Id: Ia1352f6ad170d9936113bcb4bd33fada85034dd0