Update node-feature-discovery helm chart to current version
The existing helm chart for "node feature discovery" seems to be
out of date and doesn't run properly.  Updating it to the current
version (v0.8.1) seems to fix things.

The upstream code now has a helm chart, so I've aligned with it
to simplify porting in the future.

I made one change from the upstream code, I set the chart version to
0.8.1 to match the git tag and chart appVersion.  This ensures that the
chart tarball is versioned appropriately.

At some point we could modify the build system for this chart to
use a "git repo + patches" approach rather than duplicating the
chart like I've done here.  This is out of scope for this commit
however.

Closes-Bug: 1927289
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: I8d9b6830a3852ae269b70c0e9d42ea8abf4e9458