Add helm chart for observing secrets
This commit adds helm chart for observing changes to a kubernetes
secret. This is done with a cron job. The cron job updates a
configurable deployment annotation with a checksum of the secret.
This way, when the secret changes, the pod in the deployment is
automatically restarted to pick up the changes.

Change-Id: I1a1a1b78ec1c6752747ea2a115ecea9caffdfb66
Story: 2007361
Task: 42932
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>