Include CONTAINER env var in fm helm chart
Currently the stx-openstack application apply is failing during
the fm-rest-api helm chart, and the pod log indicate a file not
found error while creating the log handler, as it uses a syslog
handler (/dev/log socket) that doesn't exist on the pod.

This change includes an environment variable to indicate that
it is running in a container, so that when in this condition fm
writes it's logs to standard output instead of using the syslog
handler.

Partial-Bug: 1951579
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>
Change-Id: Icd853ae34851df8878843730143bb1e5f8d6a6fd