Add liveness probe to fm-rest-api
The goal of this change is to use the fm-rest-api TCP port to
check if the service is listening, thus ensuring that the
application has at least started correcty.

Test Plan: Assure that the liveness probe is working as expected.

PASS: stx-openstack apply with liveness probe for fm-rest-api.
FAIL (expected): Liviness probes identify problem when fm-rest-api
runs on an unexpected TCP port.

Story: 2009745
Task: 44191

Signed-off-by: Maik Catrinque <maik.wandercatrinqueandrade@windriver.com>
Change-Id: I83ee1138ffc715127ffba1476bfee8029d418939