Add resources specification to fm-rest-api
Some pods are being evicted due to some containers
exceeding the usage of a specific resource. The goal
of this change is to be able to specify values of
limits and requests for fm-rest-api pods. In this
way we can make sure that the system has the necessary
resources to support the api.

Test Plan:

PASS: Check that with the 'enabled' flag set to 'false' no
      values of requests and limits are specified.
PASS: Check that with the 'enabled' flag set to 'true' the
      default values of requests and limits takes place.
PASS: Override the default value of requests/limits and set the
      'enabled' flag to 'true' and check that the new value
      takes place in the description of the pod.

Partial-bug: 1970645

Signed-off-by: Rafael Falcao <rafael.vieirafalcao@windriver.com>
Change-Id: I8a247c09643303f80a61b989d4b82c3835b7e601
(cherry picked from commit b34b86880113b69840ad1a12f23d0dde62b52373)