Update fm-rest-api tags to use Debian build
This change updates the fm-rest-api image tag to
'master-debian-stable-latest', due to the migration of fm-rest-api
docker image to Debian.

Changing the tag to "master-debian-stable-latest", it will be replaced
by Jenkins (build-helm-charts.sh) with the timestamped tag of the
current build.

Test plan:
PASS: Apply openstack application and update the fm-rest-api url to pull
this new image tag. Verify that:
    - pods are up and running with the new image/tag specified.
    - the container is running on Debian.
    - from inside the container, fm querys are working as expected.
    - server listening on port 18003 from inside the container
    is working as expected.

Story: 2009831
Task: 46721

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I981dc8eff7906942719222aabb9ff11b974bb01a