Update node-feature-discovery(NFD) helm chart to current version
The existing helm chart for "node feature discovery" is out of date.
This change updates it to the current version (v0.12.1).

There is one difference from upstream code: Chart version set to 0.12.1
to match the git tag and chart appVersion. This ensures that the chart
tarball is versioned appropriately.

Test plan:
[PASS] Run NFD helm chart.
[PASS] Verify NFD pods and labels are created successfully.

Signed-off-by: Davi Frossard <dbarrosf@windriver.com>
Change-Id: I5ec5ee4c8a9439a4445e7256fc19c6ac27ee9078