Update curl image version
The curl image version, used by secret-observer, was updated to 8.4.0.
The secret-observer chart version was updated to 0.1.1.

PASS - On an AIO-SX deploy, with the new OIDC app running and working,
1) edit secret oidc-auth-apps-certificate in the field "data:tls.crt"
and check that oidc-dex-*** and stx-oidc-client-*** pods are restarted
within 15 minutes; 2) edit secret dex-ca-cert in the field
"data:dex-ca-cert.crt" and check that stx-oidc-client-*** pods are
restarted within 15 minutes.

Story: 2010703
Task: 49087

Change-Id: Ie1439022601e3f96b22a77802e7ccfb73df8fdba
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>