Fixing duplicate mount path error.
When we have multiple files to observe from the
same secret and we have the same deployment to
restart for any change in file/secret, it results
in duplicate mount path error.

Test Plan:
PASSED: Freshly built the Oidc app with this change and deployed.
        Application upload/apply is successful. Verified Oidc
        pods are up and running.
PASSED: Configured the Secret-observer to run in every 1 minute.
        Secret-observer pod is created, and job is completed
        as per the scheduled interval.

Story: 2010604
Task: 49989

Change-Id: Icc2a86cdf92e3c3231ef8c32686ee8406f32bccd
Signed-off-by: Sabyasachi Nayak <sabyasachi.nayak@windriver.com>