OIDC App Up Versioning
The objective of this commit is to up version the oidc app to use the
latest stable image versions of dex, dex helm chart and curl.

Images           From        To
-------------    -------     -------
dex              0.15.3      0.18.0
helm-charts/dex  v2.37.0     v2.40.0
curl             8.4.0       8.8.0

The folowing test plan were executed for each supported kubernetes
version: v1.29.2, v1.28.4, v1.27.5, v1.26.1, v1.25.3, v1.24.4.

Test Plan:
PASS: Build an ISO successfully.
PASS: Deploy a SX and check if the new version is 24.09-54 using
      system application-list command (oidc-auth-apps 24.09-54)
PASS: Initial app install. Apply the app that is present in the
      deployed system, the oidc-auth-apps should be successfully
      installed and on applied state
PASS: Test that the new app is working with a functional test by
      creating user using manage_local_ldap_account.yml ansible
      playbook, login with this new user, using the oidc-auth
      command, authenticate with oidc-auth-apps in order to
      provide a valid token to it. Observe that the user is now
      able to successfully request kubernetes API.
PASS: Update the oidc certificate and check if oidc pods and
      secret-observer pods were successfully restarted, once its
      done, authenticate a user using oidc-auth, check if the
      user received a valid token and if it are able to do
      sucessfully requests to the kubernetes API.

Story: 2011085
Task: 50458

Change-Id: I8bbc49b14dc7de657c58ca1eb8e7e48052298fa8
Signed-off-by: Joaci Morais <Joaci.deMorais@windriver.com>