Add new label to fm-rest-api pods
Add the label 'app.starlingx.io/component' to all
fm-rest-api pods.

The value of the label can only be 'platform' or 'application', if
the value is different from that, the label will not be applied.

The value of 'app.starlingx.io/component' will be set by the
label 'isApplication' in the values file, 'true' for 'application' and
'false' for 'platform'.

By default, all pods start with the value 'platform'.

Test Plan:
PASS - The Helm charts are successfully built
PASS - The pods are deployed with label value set to platform
PASS - The pods are restarted if the value changes

Story: 2010612
Task: 50517

Change-Id: Ib8c5826c6f5630c1180b3765ba6037150a8ae705
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>