Add new ttl value to fm-rest-api values.yaml
STX-Openstack apply was failing due to a low value of a ttl
set to the jobs. Some jobs were being cleaned before the
deployment of the Helm chart was concluded.

This review expose the configuration of the ttl to the values
file of the fm-rest-api Helm chart.

Test Plan:
PASS - fm-rest-api is deployed with the value set in the values file

Partial-Bug: 2075509

Change-Id: Ied933ef360342f9eb200d56035a27bf0299b503c
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>