Remove ttl from fm-rest-api values
This review removes the ttlSecondsAfterFinished from the
values.yaml of fm-rest-api Helm Chart.

The deletion of the jobs using the ttlSecondsAfterFinished
was causing some pods of the application to be stuck in
init state after a host reboot.

Test Plan:
PASS: fm-rest-api package is built
PASS: None of the jobs are deleted
PASS: All fm-rest-api pods come back after host reboot

Depends-On: https://review.opendev.org/c/starlingx/openstack-armada-app/+/925972

Change-Id: Ibf2c494e54280f976fa8bece8d05856ae50d1d12
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>