Add pre-apply cleanup Job to fm-rest-api
After verification, it was noted that it is not possible
to reapply STX-Openstack after a helm-override that changes
a job template since the template section of a job is
immutable or not updatable

Due to the use of kubernetes-entrypoint DEPENDENCY_JOBS it was
also noted that deleting the jobs after the application is applied
it is not an option. If this happens, the application would not
come back after a host reboot.

This review creates a Job that runs right before the Helm chart is
installed or updated. This Job deletes all jobs that have its status
as completed.

Depends-On: https://review.opendev.org/c/starlingx/openstack-armada-app/+/926561

Test Plan:
PASS - Build fm-rest-api
PASS - Deploy fm-rest-api
PASS - Reconciliate after values change to fm-rest-api
PASS - fm-rest-api pods are up after host reboot

Change-Id: I4d35a1fae339aa6d78d9076065d2613f76f30c69
Signed-off-by: Daniel Caires <DanielMarques.Caires@windriver.com>